requirements.txt
--extra-index-url https://test.pypi.org/simple/
docker~=4.4.4
flask~=1.1.2
requests~=2.25.1
chess~=1.3.3
cuwais-common~=1.2.2-dev2106051
SQLAlchemy~=1.4.3
psycopg2~=2.8.6
python-socketio
numpy~=1.20.2
Werkzeug~=1.0.1
eventlet~=0.31.0
gunicorn
